+++ packages/terra-dialog-modal/src/terra-dev-site/test/dialog-modal/DialogModalWithLongText.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Button from 'terra-button';
 import ActionHeader from 'terra-action-header';
 import ActionFooter from 'terra-action-footer';
-import DialogModal from '../../../../src/DialogModal';
+import DialogModal from '../../../DialogModal';
 
 
 class DialogModalWithLongText extends React.Component {